+++ lib/hashtable.hh
@@ -315,8 +315,17 @@ public:
         : Hashtable(ksize, new ByteStorage(sizes)) { } ;
 };
 
+// Hashtable-derived class with NibbleStorage.
+class SmallCounttable : public khmer::Hashtable
+{
+public:
+    explicit SmallCounttable(WordLength ksize, std::vector<uint64_t> sizes)
+        : Hashtable(ksize, new NibbleStorage(sizes)) { } ;
+};
+
 // Hashtable-derived class with BitStorage.
-class Nodetable : public Hashtable {
+class Nodetable : public Hashtable
+{
 public:
     explicit Nodetable(WordLength ksize, std::vector<uint64_t> sizes)
         : Hashtable(ksize, new BitStorage(sizes)) { } ;