+++ tests/test_script_arguments.py
@@ -151,6 +151,7 @@ FakeArgparseObject = collections.namedtuple('FakeArgs',
                                              'max_tablesize',
                                              'max_memory_usage',
                                              'unique_kmers',
+                                             'small_count',
                                              'force'])
 
 
@@ -160,7 +161,8 @@ def test_create_countgraph_1():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     countgraph = khmer_args.create_countgraph(args)
     expected_hashsz = utils.longify([2499997, 2499989, 2499983, 2499967])
@@ -176,7 +178,8 @@ def test_create_countgraph_2():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     countgraph = khmer_args.create_countgraph(args, ksize=15)
     assert countgraph.ksize() == 15
@@ -190,7 +193,8 @@ def test_create_countgraph_3():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     old_stderr = sys.stderr
     sys.stderr = capture = StringIO()
@@ -213,7 +217,8 @@ def test_create_countgraph_4():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     old_stderr = sys.stderr
     sys.stderr = capture = StringIO()
@@ -236,7 +241,8 @@ def test_create_countgraph_5():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 1)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 1)
 
     old_stderr = sys.stderr
     sys.stderr = capture = StringIO()
@@ -258,7 +264,8 @@ def test_create_countgraph_4_multiplier():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     countgraph = khmer_args.create_countgraph(args, multiplier=2.0)
     assert sum(countgraph.hashsizes()) < max_mem / 2.0, \
@@ -271,7 +278,8 @@ def test_create_nodegraph_1():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     nodegraph = khmer_args.create_nodegraph(args)
     expected_hashsz = utils.longify([19999999, 19999981, 19999963, 19999927])
@@ -289,7 +297,8 @@ def test_create_nodegraph_2():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     nodegraph = khmer_args.create_nodegraph(args, ksize=15)
     assert nodegraph.ksize() == 15
@@ -303,7 +312,8 @@ def test_create_nodegraph_3():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     sys.stderr = capture = StringIO()
 
@@ -323,7 +333,8 @@ def test_create_nodegraph_4():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     sys.stderr = capture = StringIO()
 
@@ -343,7 +354,8 @@ def test_create_nodegraph_5():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 1)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 1)
 
     sys.stderr = capture = StringIO()
 
@@ -362,7 +374,8 @@ def test_create_nodegraph_4_multiplier():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     nodegraph = khmer_args.create_nodegraph(args, multiplier=2.0)
     assert sum(nodegraph.hashsizes()) / 8.0 < max_mem / 2.0, \
@@ -375,7 +388,8 @@ def test_report_on_config_bad_graphtype():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     try:
         khmer_args.report_on_config(args, 'foograph')
@@ -392,7 +406,8 @@ def test_fail_calculate_foograph_size():
     max_tablesize = khmer_args.DEFAULT_MAX_TABLESIZE
     max_mem = 1e7
 
-    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0, 0)
+    args = FakeArgparseObject(ksize, n_tables, max_tablesize, max_mem, 0,
+                              False, 0)
 
     try:
         khmer_args.calculate_graphsize(args, 'foograph')